// React/app/GerAdv_TS/Historico/Interfaces/interface.Historico.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IHistorico {
// 202501251
    id: number;
 
	processo: number,
	precatoria: number,
	apenso: number,
	fase: number,
	statusandamento: number,
	extraid: number,
	idne: number,
	extraguid: string,
	liminarorigem: number,
	naopublicavel: boolean,
	idinstprocesso: number,
	data: string,
	observacao: string,
	agendado: boolean,
	concluido: boolean,
	mesmaagenda: boolean,
	sad: number,
	resumido: boolean,
	top: boolean,
}

export interface IHistoricoFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IHistoricoIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}